# docs/architecture/05-api-contracts.yaml
openapi: 3.1.0
info:
  title: Signature Router & Management System API
  version: 1.0.0
  description: |
    Banking-grade digital signature orchestration system.
    Supports multiple channels: SMS, Push, Voice, Biometric.
    
    ## Features
    - Dynamic routing with SpEL rules
    - Automatic fallback on failure
    - Provider degraded mode handling
    - Complete audit trail
    - Idempotency support
    
    ## SLOs
    - P99 latency: < 300ms
    - Availability: ≥ 99.9%
    - Error rate: < 0.1%
    
  contact:
    name: Banking Platform Team
    email: platform@bank.com
servers:
  - url: https://api.bank.com/signature-router/v1
    description: Production
  - url: https://api-staging.bank.com/signature-router/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /admin/rules:
    get:
      tags: [Admin - Rules]
      summary: List routing rules
      description: Get all routing rules ordered by priority
      operationId: listRules
      security:
        - BearerAuth: [admin]
      
      parameters:
        - name: enabled
          in: query
          schema:
            type: boolean
          description: Filter by enabled status
      
      responses:
        '200':
          description: List of routing rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoutingRuleResponse'
              example:
                - id: "018c7f5e-rule-7890-abcd-ef1234567890"
                  name: "High Risk Transactions"
                  description: "High risk always via SMS"
                  priority: 10
                  condition: "context.riskLevel == 'HIGH'"
                  targetChannel: "SMS"
                  enabled: true
                  createdAt: "2025-11-01T00:00:00Z"
                  createdBy: "admin@bank.com"
                - id: "018c7f5e-rule-7890-abcd-ef1234567891"
                  name: "Large Amounts"
                  description: "Amounts > 10K via Voice"
                  priority: 20
                  condition: "context.amount.value > 10000 && context.amount.currency == 'USD'"
                  targetChannel: "VOICE"
                  enabled: true
                  createdAt: "2025-11-01T00:00:00Z"
                  createdBy: "admin@bank.com"
        
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    
    post:
      tags: [Admin - Rules]
      summary: Create routing rule
      description: |
        Create new routing rule with SpEL condition.
        
        **SpEL Context Available**:
        - `context.transactionId` (String)
        - `context.transactionType` (String)
        - `context.amount.value` (BigDecimal)
        - `context.amount.currency` (String)
        - `context.riskLevel` (String: HIGH, MEDIUM, LOW)
        - `context.customProperty` (any JSONB field)
        
        **Examples**:
        - `context.riskLevel == 'HIGH'`
        - `context.amount.value > 10000`
        - `context.transactionType == 'WIRE_TRANSFER'`
        - `context.amount.value > 10000 && context.riskLevel == 'HIGH'`
      
      operationId: createRule
      security:
        - BearerAuth: [admin]
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleRequest'
            example:
              name: "International Wires"
              description: "International wire transfers via Voice for extra security"
              priority: 15
              condition: "context.transactionType == 'WIRE_TRANSFER' && context.destination == 'INTERNATIONAL'"
              targetChannel: "VOICE"
              enabled: true
      
      responses:
        '201':
          description: Rule created successfully
          headers:
            Location:
              schema:
                type: string
              example: "/api/v1/admin/rules/018c7f5e-rule-7890-abcd-ef1234567890"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingRuleResponse'
        
        '400':
          description: Invalid SpEL expression
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "INVALID_SPEL"
                message: "SpEL expression syntax error"
                details:
                  - field: "condition"
                    error: "Parse error at position 15: unexpected token 'xyx'"
                timestamp: "2025-11-26T10:30:00Z"
                traceId: "trace-abc123"
        
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/rules/{id}:
    get:
      tags: [Admin - Rules]
      summary: Get routing rule
      operationId: getRule
      security:
        - BearerAuth: [admin]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Rule found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingRuleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags: [Admin - Rules]
      summary: Update routing rule
      operationId: updateRule
      security:
        - BearerAuth: [admin]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuleRequest'
      responses:
        '200':
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingRuleResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags: [Admin - Rules]
      summary: Delete routing rule
      operationId: deleteRule
      security:
        - BearerAuth: [admin]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Rule deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # ADMIN - CONNECTORS
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource ID (UUIDv7)
  schemas:
...
    CreateRuleRequest:
      type: object
      required:
        - name
        - priority
        - condition
        - targetChannel
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        priority:
          type: integer
          minimum: 0
          description: Lower = higher priority (first match wins)
        condition:
          type: string
          description: SpEL expression evaluated against TransactionContext
          example: "context.riskLevel == 'HIGH'"
        targetChannel:
          type: string
          enum: [SMS, PUSH, VOICE, BIOMETRIC]
        enabled:
          type: boolean
          default: true
    UpdateRuleRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
          minimum: 0
        condition:
          type: string
        targetChannel:
          type: string
          enum: [SMS, PUSH, VOICE, BIOMETRIC]
        enabled:
          type: boolean
    RoutingRuleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
        condition:
          type: string
        targetChannel:
          type: string
          enum: [SMS, PUSH, VOICE, BIOMETRIC]
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedBy:
          type: string
    
    # ========================================
    # Connector Schemas
    # ========================================
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, NOT_FOUND, UNAUTHORIZED, FORBIDDEN, INTERNAL_ERROR, SERVICE_UNAVAILABLE, INVALID_SPEL]
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              error:
                type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          description: Distributed trace ID for debugging
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "UNAUTHORIZED"
            message: "Missing or invalid authentication token"
            timestamp: "2025-11-26T10:30:00Z"
            traceId: "trace-abc123"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "FORBIDDEN"
            message: "Insufficient permissions for this operation"
            timestamp: "2025-11-26T10:30:00Z"
            traceId: "trace-abc123"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "NOT_FOUND"
            message: "Signature request not found"
            timestamp: "2025-11-26T10:30:00Z"
            traceId: "trace-abc123"
...